Combined UrlHeaderservice into urlactionservice
