+++ Sulfur/Controllers/TorrentFileController.cs
@@ -8,7 +8,6 @@ using Sulfur.Constants;
 using Sulfur.Models;
 using Sulfur.Models.Db;
 using Sulfur.Services.UrlHeaderActions;
-using Sulfur.Services.UrlPayloadActions;
 
 namespace Sulfur
 {
@@ -17,14 +16,12 @@ namespace Sulfur
     public class TorrentFileController : ControllerBase
     {
         private readonly IDbContext _context;
-        private readonly IUrlHeaderService _urlHeaderService;
-        private readonly IUrlPayloadService _urlPayloadService;
+        private readonly IUrlAactionService _urlActionService;
 
-        public TorrentFileController(IDbContext context, IUrlPayloadService urlPayloadService, IUrlHeaderService urlHeaderService)
+        public TorrentFileController(IDbContext context, IUrlAactionService urlActionService)
         {
             _context = context;
-            _urlPayloadService = urlPayloadService;
-            _urlHeaderService = urlHeaderService;
+            _urlActionService = urlActionService;
         }
 
         // POST api/TorrentFile
@@ -37,14 +34,14 @@ namespace Sulfur
 
             //Checks if it is possible to bind the values in the request to the model.
             //If url is null, means it was unable to bind the model
-            if (!ModelState.IsValid || !_urlHeaderService.SuccessfulMatchOnHeaderToken(headerAuthToken))
+            if (!ModelState.IsValid || !_urlActionService.SuccessfulMatchOnHeaderToken(headerAuthToken))
             {
                 //Returns a 400 bad request
                 return BadRequest(ModelState);
             }
 
             //Check also if the authtoken is valid also
-            if (!_urlHeaderService.SuccessfulMatchOnHeaderToken(headerAuthToken))
+            if (!_urlActionService.SuccessfulMatchOnHeaderToken(headerAuthToken))
             {
                 //Returns a 400 bad request
                 return BadRequest();
@@ -53,7 +50,7 @@ namespace Sulfur
             //ActionResult is the base class for various results for example JSONResult or Result
             //You can return a various amount of things.
             //Return a guid value from the post request
-            return _urlPayloadService.GenerateGuidPayload();
+            return _urlActionService.GenerateGuidPayload();
         }
     }
 }
