--- Sulfur/Services/UrlHeaderActions/UrlHeaderService.cs
@@ -1,17 +1,0 @@
-﻿using Sulfur.Constants;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Sulfur.Services.UrlHeaderActions
-{
-    public class UrlHeaderService : IUrlHeaderService
-    {
-        //Checks whether the auth header token matches our predefined value
-        public bool SuccessfulMatchOnHeaderToken(string headerToken)
-        {
-            return headerToken.Equals(ServiceConstants.AuthTokenPassword);
-        }
-    }
-}
