--- Sulfur/Services/UrlPayloadActions/UrlPayloadService.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Sulfur.Models;
-
-namespace Sulfur.Services.UrlPayloadActions
-{
-    public class UrlPayloadService : IUrlPayloadService
-    {
-        //Create a new Guid code to be sent back as a response.
-        //TODO In the future, the UrlPayload will be 
-        GuidResult IUrlPayloadService.GenerateGuidPayload()
-        {
-            Guid code = Guid.NewGuid();
-            GuidResult guidCode = new GuidResult();
-            guidCode.Id = code.ToString();
-            return guidCode;
-        }
-    }
-}
